move prompts into utils
