+++ packages/expo-cli/src/credentials/actions/promptForCredentials.ts
@@ -4,7 +4,7 @@ import path from 'path';
 import untildify from 'untildify';
 
 import Log from '../../log';
-import prompts, { Question as PromptQuestion } from '../../prompts';
+import prompts, { Question as PromptQuestion } from '../../utils/prompts';
 import * as validators from '../../utils/validators';
 
 export type Question = {
