+++ packages/expo-cli/src/credentials/views/IosPushCredentials.ts
@@ -5,8 +5,8 @@ import terminalLink from 'terminal-link';
 import CommandError from '../../CommandError';
 import { isPushKey, PushKey, PushKeyInfo, PushKeyManager } from '../../appleApi';
 import Log from '../../log';
-import prompt, { confirmAsync, Question } from '../../prompts';
 import { ora } from '../../utils/ora';
+import prompt, { confirmAsync, Question } from '../../utils/prompts';
 import { displayIosUserCredentials } from '../actions/list';
 import { askForUserProvided, CredentialSchema } from '../actions/promptForCredentials';
 import { AppLookupParams, getAppLookupParams } from '../api/IosApi';
